api/requirements.txt
azure-functions
pydantic>=2.0.0
python-jose[cryptography]
bcrypt
python-multipart
psycopg2-binary
sqlalchemy>=2.0.0
alembic
python-dotenv
